requirements.txt
opencv-python
moviepy
pillow
gtts